+++ src/scripts/InstanceScripts/Instance_UtgardeKeep.cpp
@@ -421,7 +421,7 @@ class SkarvaldTheConstructorAI : public CreatureAIScript
                 sendChatMessage(CHAT_MSG_MONSTER_YELL, 0, "Not... over... yet.");
                 pDalronn->sendChatMessage(CHAT_MSG_MONSTER_YELL, 13203, "Skarvald, you incompetent slug! Return and make yourself useful!");
                 spawnCreature(CN_SKARVALD_GHOST, getCreature()->GetPosition());
-                getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
+                getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
             }
             else if (pDalronn != nullptr && !pDalronn->isAlive())
             {
@@ -504,7 +504,7 @@ class DalronnTheControllerAI : public CreatureAIScript
                 sendChatMessage(CHAT_MSG_MONSTER_YELL, 0, "See... you... soon.");
                 pSkarvald->sendChatMessage(CHAT_MSG_MONSTER_YELL, 13233, "Pagh! What sort of necromancer lets death stop him? I knew you were worthless!");
                 spawnCreature(CN_DALRONN_GHOST, getCreature()->GetPosition());
-                getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
+                getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
             }
             else if (pSkarvald != nullptr && !pSkarvald->isAlive())
             {
@@ -554,7 +554,7 @@ class SkarvaldTheConstructorGhostAI : public CreatureAIScript
 
         void OnLoad() override
         {
-            getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_IGNORE_PLAYER_COMBAT);
+            getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_IGNORE_PLAYER_COMBAT);
 
             Player* pTarget = getNearestPlayer();
             if (pTarget != nullptr)
@@ -581,7 +581,7 @@ class DalronnTheControllerGhostAI : public CreatureAIScript
 
         void OnLoad() override
         {
-            getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_IGNORE_PLAYER_COMBAT);
+            getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_IGNORE_PLAYER_COMBAT);
 
             Player* pTarget = getNearestPlayer();
             if (pTarget != nullptr)
