+++ src/scripts/InstanceScripts/Raid_TheObsidianSanctum.cpp
@@ -318,7 +318,7 @@ class TsunamiAI : public CreatureAIScript
             RegisterAIUpdateEvent(1000);
             setFlyMode(true);
             setCanEnterCombat(false);
-            getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
+            getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
             despawn(11500, 0);
         }
 
@@ -341,7 +341,7 @@ class CyclonAI : public CreatureAIScript
         {
             setRooted(true);
             setCanEnterCombat(false);
-            getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
+            getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
             _applyAura(CYCLON_SPELL);
             _applyAura(CYCLON_AURA);
         }
