+++ src/scripts/SpellHandlers/QIspells.cpp
@@ -1483,7 +1483,7 @@ bool ToLegionHold(uint8_t /*effectIndex*/, Aura* pAura, bool apply)
         Creature* pJovaan = pPlayer->GetMapMgr()->CreateAndSpawnCreature(21633, -3310.743896f, 2951.929199f, 171.132538f, 5.054039f);    // Spawn Jovaan
         if (pJovaan != nullptr)
         {
-            pJovaan->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_ATTACKABLE_2);
+            pJovaan->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_ATTACKABLE_2);
             if (pJovaan->GetAIInterface() != nullptr)
             {
                 pJovaan->GetAIInterface()->SetAllowedToEnterCombat(false);
